smartxml: Allow Fragments to refresh their node references using their initial paths
This allows client code to refresh Fragment object to reference
proper node instances after some potentially invalidating
changes to a document has been made while the logical structure
of the document is known to be the same.

An example of such a scenario would be a fragment instance created
before performing a transaction that was later undone using full
document reload.