smartxml: Cache DocumentNode instances in the underlying DOM structure
This allows for data persistence in the DocumentNode.object api. And,
at last, puts an end to the creation of unnecessary instances.

The side effect is that extensions needs to be registered before
acquiring references to nodes as they won't be updated with extension
behavior after they are created. This probably should be better
reflected in the api, maybe by setting extension via document constructor.